Fix a use-after-free.
CString owns the memory for the string it creates, and so when it goes
out of scope, it will drop the memory. This means that the old usage of
it hands out a dangling pointer, which led to use-after-free. This code
isn't super clean, but it does the job for now.