Code Triage badge
Code Triage allows people to receive emails of random project issues in their email daily to help them motivate themselves into contributing and helping the project advance.  This badge shows the number of people who've subscribed to the issue mailings.  Note that some people may just be interested in the issues themselves rather than contributing.  Still this is an ideal way to volunteer.